Move ftp exception to new package
